Add video studio template handlers
